Reorganize src/ into base/, interp/, hdf5/ subfolders
Move 26 headers from flat src/ + src/detail/ into three logical
subfolders: base/ (5 files), interp/ (10 files), hdf5/ (11 files).

- Convert all #include paths to include-root-relative form
- Add explicit includes to HDF5 loader files that relied on umbrella ordering
- Update test and Cactus thorn includes to use new paths
- Remove src/detail/ directory
