--- src/WeakLibReader_Hdf5Loader.hpp
@@ -1,21 +1,0 @@
-#pragma once
-
-#include <AMReX_Array.H>
-#include <AMReX_GpuContainers.H>
-#include <AMReX_ParallelDescriptor.H>
-
-#include <limits>
-#include <string>
-#include <utility>
-
-#include "WeakLibReader_Hdf5Types.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderDetail.hpp"
-
-#include "detail/WeakLibReader_Hdf5LoaderTable.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderEos.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderOpacityEmAb.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderOpacityScat.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderOpacityMain.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderOpacityDevice.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderParallel.hpp"
-#include "detail/WeakLibReader_Hdf5LoaderOpacityParallel.hpp"
