--- src/WeakLibReader_InterpLogTable.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "detail/WeakLibReader_InterpLogTablePoint.hpp"
-#include "detail/WeakLibReader_InterpLogTableDeriv.hpp"
-#include "detail/WeakLibReader_InterpLogTableSlice.hpp"
