--- src/WeakLibReader_LogInterpolate.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "detail/WeakLibReader_LogInterpolateCore.hpp"
-#include "detail/WeakLibReader_LogInterpolatePoint.hpp"
-#include "detail/WeakLibReader_LogInterpolateSweep.hpp"
-#include "detail/WeakLibReader_LogInterpolateDeriv.hpp"
-#include "detail/WeakLibReader_LogInterpolateSum.hpp"
