+++ src/base/WeakLibReader_InterpBasis.hpp
