+++ src/base/WeakLibReader_Math.hpp
