+++ src/hdf5/WeakLibReader_Hdf5Loader.hpp
@@ -1,0 +1,21 @@
+#pragma once
+
+#include <AMReX_Array.H>
+#include <AMReX_GpuContainers.H>
+#include <AMReX_ParallelDescriptor.H>
+
+#include <limits>
+#include <string>
+#include <utility>
+
+#include "hdf5/WeakLibReader_Hdf5Types.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderDetail.hpp"
+
+#include "hdf5/WeakLibReader_Hdf5LoaderTable.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderEos.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityEmAb.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityScat.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityMain.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityDevice.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderParallel.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityParallel.hpp"
