+++ src/hdf5/WeakLibReader_Hdf5LoaderOpacityEmAb.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "hdf5/WeakLibReader_Hdf5Types.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderDetail.hpp"
+
 namespace WeakLibReader {
 namespace detail {
