+++ src/hdf5/WeakLibReader_Hdf5LoaderOpacityMain.hpp
@@ -1,5 +1,10 @@
 #pragma once
 
+#include "hdf5/WeakLibReader_Hdf5Types.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderDetail.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityEmAb.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityScat.hpp"
+
 namespace WeakLibReader {
 namespace detail {
