+++ src/hdf5/WeakLibReader_Hdf5LoaderOpacityParallel.hpp
@@ -1,5 +1,9 @@
 #pragma once
 
+#include "hdf5/WeakLibReader_Hdf5Types.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderDetail.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderOpacityMain.hpp"
+
 namespace WeakLibReader {
 namespace detail {
