+++ src/hdf5/WeakLibReader_Hdf5LoaderTable.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "hdf5/WeakLibReader_Hdf5Types.hpp"
+#include "hdf5/WeakLibReader_Hdf5LoaderDetail.hpp"
+
 namespace WeakLibReader {
 inline Hdf5LoadStatus LoadHdf5Table(const std::string& filePath,
                                     Hdf5Table& output,
