+++ src/interp/WeakLibReader_InterpLogTable.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "interp/WeakLibReader_InterpLogTablePoint.hpp"
+#include "interp/WeakLibReader_InterpLogTableDeriv.hpp"
+#include "interp/WeakLibReader_InterpLogTableSlice.hpp"
