+++ src/interp/WeakLibReader_LogInterpolate.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "interp/WeakLibReader_LogInterpolateCore.hpp"
+#include "interp/WeakLibReader_LogInterpolatePoint.hpp"
+#include "interp/WeakLibReader_LogInterpolateSweep.hpp"
+#include "interp/WeakLibReader_LogInterpolateDeriv.hpp"
+#include "interp/WeakLibReader_LogInterpolateSum.hpp"
