+++ src/interp/WeakLibReader_LogInterpolateSum.hpp
@@ -4,7 +4,7 @@
 #include <cstddef>
 #include <vector>
 
-#include "WeakLibReader_LogInterpolateCore.hpp"
+#include "interp/WeakLibReader_LogInterpolateCore.hpp"
 
 namespace WeakLibReader {
