+++ test/test_hdf5_loader.cpp
@@ -1,7 +1,7 @@
 #define SIMPLE_CATCH_NO_MAIN
 #include <catch2/catch_test_macros.hpp>
 
-#include "WeakLibReader_Hdf5Loader.hpp"
+#include "hdf5/WeakLibReader_Hdf5Loader.hpp"
 
 #include <AMReX.H>
 #include <AMReX_GpuContainers.H>
