+++ test/test_log_interpolate_3d.cpp
@@ -1,9 +1,9 @@
 #define SIMPLE_CATCH_NO_MAIN
 #include <catch2/catch_test_macros.hpp>
 
-#include "WeakLibReader_LogInterpolate.hpp"
-#include "WeakLibReader_Layout.hpp"
-#include "WeakLibReader_AxisTypes.hpp"
+#include "interp/WeakLibReader_LogInterpolate.hpp"
+#include "base/WeakLibReader_Layout.hpp"
+#include "base/WeakLibReader_AxisTypes.hpp"
 
 #include <array>
 #include <cmath>
