+++ test/test_log_interpolate_basis.cpp
@@ -1,7 +1,7 @@
 #define SIMPLE_CATCH_NO_MAIN
 #include <catch2/catch_test_macros.hpp>
 
-#include "WeakLibReader_InterpBasis.hpp"
+#include "base/WeakLibReader_InterpBasis.hpp"
 
 #include <cmath>
